+++ src/defaults.ts
@@ -22,7 +22,7 @@ export const horizons = ['0%'] as const ;
 export const mode = 'automatic';
 export const resolveBareModules = true;
 export const forceCleanNpmInstall = false;
-export const globalMeasurementExpression = 'window.tachometerResult';
+export const measurementExpression = 'window.tachometerResult';
 
 export function measurement(url: LocalUrl|RemoteUrl): Measurement {
   if (url.kind === 'remote') {